Moved API requests to a top level directory
