+++ apps/admin-x-settings/src/api/themes.ts
@@ -1,5 +1,5 @@
-import {InstalledTheme, Theme} from '../../types/api';
-import {createMutation, createQuery} from '../apiRequests';
+import {InstalledTheme, Theme} from '../types/api';
+import {createMutation, createQuery} from '../utils/apiRequests';
 
 export interface ThemesResponseType {
     themes: Theme[];
