+++ apps/admin-x-settings/src/api/tiers.ts
@@ -1,5 +1,5 @@
-import {Meta, createMutation, createQuery} from '../apiRequests';
-import {Tier} from '../../types/api';
+import {Meta, createMutation, createQuery} from '../utils/apiRequests';
+import {Tier} from '../types/api';
 
 export interface TiersResponseType {
     meta?: Meta
