+++ apps/admin-x-settings/src/components/settings/general/Facebook.tsx
@@ -5,7 +5,7 @@ import SettingGroupContent from '../../../admin-x-ds/settings/SettingGroupConten
 import TextField from '../../../admin-x-ds/global/form/TextField';
 import useSettingGroup from '../../../hooks/useSettingGroup';
 import {ReactComponent as FacebookLogo} from '../../../admin-x-ds/assets/images/facebook-logo.svg';
-import {getImageUrl, useUploadImage} from '../../../utils/api/images';
+import {getImageUrl, useUploadImage} from '../../../api/images';
 import {getSettingValues} from '../../../utils/helpers';
 
 const Facebook: React.FC<{ keywords: string[] }> = ({keywords}) => {
