+++ apps/admin-x-settings/src/components/settings/general/InviteUserModal.tsx
@@ -5,8 +5,8 @@ import TextField from '../../../admin-x-ds/global/form/TextField';
 import useRouting from '../../../hooks/useRouting';
 import validator from 'validator';
 import {showToast} from '../../../admin-x-ds/global/Toast';
-import {useAddInvite} from '../../../utils/api/invites';
-import {useBrowseRoles} from '../../../utils/api/roles';
+import {useAddInvite} from '../../../api/invites';
+import {useBrowseRoles} from '../../../api/roles';
 import {useEffect, useRef, useState} from 'react';
 
 type RoleType = 'administrator' | 'editor' | 'author' | 'contributor';
