+++ apps/admin-x-settings/src/components/settings/general/Twitter.tsx
@@ -5,7 +5,7 @@ import SettingGroupContent from '../../../admin-x-ds/settings/SettingGroupConten
 import TextField from '../../../admin-x-ds/global/form/TextField';
 import useSettingGroup from '../../../hooks/useSettingGroup';
 import {ReactComponent as TwitterLogo} from '../../../admin-x-ds/assets/images/twitter-logo.svg';
-import {getImageUrl, useUploadImage} from '../../../utils/api/images';
+import {getImageUrl, useUploadImage} from '../../../api/images';
 import {getSettingValues} from '../../../utils/helpers';
 
 const Twitter: React.FC<{ keywords: string[] }> = ({keywords}) => {
