+++ apps/admin-x-settings/src/components/settings/membership/portal/PortalLinks.tsx
@@ -6,7 +6,7 @@ import React, {useEffect, useState} from 'react';
 import Select from '../../../../admin-x-ds/global/form/Select';
 import TextField from '../../../../admin-x-ds/global/form/TextField';
 import {getHomepageUrl, getPaidActiveTiers} from '../../../../utils/helpers';
-import {useBrowseTiers} from '../../../../utils/api/tiers';
+import {useBrowseTiers} from '../../../../api/tiers';
 import {useGlobalData} from '../../../providers/GlobalDataProvider';
 
 interface PortalLinkPrefs {
