+++ apps/admin-x-settings/src/components/settings/membership/tiers/TiersList.tsx
@@ -8,7 +8,7 @@ import useRouting from '../../../../hooks/useRouting';
 import {Tier} from '../../../../types/api';
 import {currencyToDecimal, getSymbol} from '../../../../utils/currency';
 import {numberWithCommas} from '../../../../utils/helpers';
-import {useEditTier} from '../../../../utils/api/tiers';
+import {useEditTier} from '../../../../api/tiers';
 
 interface TiersListProps {
     tab?: 'active-tiers' | 'archive-tiers' | 'free-tier';
