+++ apps/admin-x-settings/src/components/settings/site/designAndBranding/BrandSettings.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 import SettingGroupContent from '../../../../admin-x-ds/settings/SettingGroupContent';
 import TextField from '../../../../admin-x-ds/global/form/TextField';
 import {SettingValue} from '../../../../types/api';
-import {getImageUrl, useUploadImage} from '../../../../utils/api/images';
+import {getImageUrl, useUploadImage} from '../../../../api/images';
 
 export interface BrandSettingValues {
     description: string
